# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ScoredHex {
  h3Index: string;
  lat: number;
  lng: number;
  demandScore: number;
  supplyScore: number;
  competitionPenalty: number;
  alphaScore: number;
  citationCount: number;
  poiCount: number;
  residentialParcelCount: number;
  publicParkingSpots: number;
  estimatedDriveways: number;
  neighborhoodName: string;
  launchZoneId: number | null;
}

export interface HexFeatureProperties {
  h3Index: string;
  alphaScore: number;
  demandScore: number;
  supplyScore: number;
  competitionPenalty: number;
  neighborhoodName: string;
  launchZoneId: number | null;
  isGoldilocksZone: boolean;
}

export interface HexFeatureGeometry {
  type: string;
  coordinates: number[][][];
}

export interface HexFeature {
  type: string;
  properties: HexFeatureProperties;
  geometry: HexFeatureGeometry;
}

export interface HexGeoJson {
  type: string;
  features: HexFeature[];
}

export interface LaunchZone {
  id: number;
  rank: number;
  centroidLat: number;
  centroidLng: number;
  hexCount: number;
  meanAlpha: number;
  maxAlpha: number;
  totalResidentialParcels: number;
  estimatedDriveways: number;
  dominantNeighborhood: string;
  label: string;
}

export interface CitySummary {
  city: string;
  totalHexes: number;
  hexesAboveThreshold: number;
  launchZoneCount: number;
  avgAlphaScore: number;
  topZoneName: string;
  topZoneScore: number;
  topHexH3: string;
  topHexAlpha: number;
  topHexLat: number;
  topHexLng: number;
  medianDemandScore: number;
  medianSupplyScore: number;
  totalEstimatedDriveways: number;
  algorithmVersion: string;
}

export type ListHexesParams = {
  minScore?: number;
};

export type GetHexGeoJsonParams = {
  layer?: string;
};