# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
servers:
  - url: /api
tags:
  - name: health
  - name: gis
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /gis/hexes:
    get:
      operationId: listHexes
      tags: [gis]
      parameters:
        - name: minScore
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: List of scored hex cells
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoredHex"
  /gis/hexes/geojson:
    get:
      operationId: getHexGeoJson
      tags: [gis]
      parameters:
        - name: layer
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: GeoJSON FeatureCollection of hex cells
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HexGeoJson"
  /gis/hexes/{h3Index}:
    get:
      operationId: getHex
      tags: [gis]
      parameters:
        - name: h3Index
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Single hex cell detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoredHex"
        "404":
          description: Not found
  /gis/summary:
    get:
      operationId: getCitySummary
      tags: [gis]
      responses:
        "200":
          description: City summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitySummary"
  /gis/launch-zones:
    get:
      operationId: listLaunchZones
      tags: [gis]
      responses:
        "200":
          description: Ranked launch zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LaunchZone"
  /gis/launch-zones/{id}:
    get:
      operationId: getLaunchZone
      tags: [gis]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Single launch zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaunchZone"
        "404":
          description: Not found
  /gis/launch-zones/{id}/hexes:
    get:
      operationId: getLaunchZoneHexes
      tags: [gis]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hexes belonging to a launch zone
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoredHex"
components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    ScoredHex:
      type: object
      required:
        - h3Index
        - lat
        - lng
        - demandScore
        - supplyScore
        - competitionPenalty
        - alphaScore
        - citationCount
        - poiCount
        - residentialParcelCount
        - publicParkingSpots
        - estimatedDriveways
        - neighborhoodName
        - launchZoneId
      properties:
        h3Index:
          type: string
        lat:
          type: number
        lng:
          type: number
        demandScore:
          type: number
        supplyScore:
          type: number
        competitionPenalty:
          type: number
        alphaScore:
          type: number
        citationCount:
          type: integer
        poiCount:
          type: integer
        residentialParcelCount:
          type: integer
        publicParkingSpots:
          type: integer
        estimatedDriveways:
          type: integer
        neighborhoodName:
          type: string
        launchZoneId:
          type: integer
          nullable: true
    HexFeatureProperties:
      type: object
      required:
        - h3Index
        - alphaScore
        - demandScore
        - supplyScore
        - competitionPenalty
        - neighborhoodName
        - launchZoneId
        - isGoldilocksZone
      properties:
        h3Index:
          type: string
        alphaScore:
          type: number
        demandScore:
          type: number
        supplyScore:
          type: number
        competitionPenalty:
          type: number
        neighborhoodName:
          type: string
        launchZoneId:
          type: integer
          nullable: true
        isGoldilocksZone:
          type: boolean
    HexFeatureGeometry:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
    HexFeature:
      type: object
      required: [type, properties, geometry]
      properties:
        type:
          type: string
        properties:
          $ref: "#/components/schemas/HexFeatureProperties"
        geometry:
          $ref: "#/components/schemas/HexFeatureGeometry"
    HexGeoJson:
      type: object
      required: [type, features]
      properties:
        type:
          type: string
        features:
          type: array
          items:
            $ref: "#/components/schemas/HexFeature"
    LaunchZone:
      type: object
      required:
        - id
        - rank
        - centroidLat
        - centroidLng
        - hexCount
        - meanAlpha
        - maxAlpha
        - totalResidentialParcels
        - estimatedDriveways
        - dominantNeighborhood
        - label
      properties:
        id:
          type: integer
        rank:
          type: integer
        centroidLat:
          type: number
        centroidLng:
          type: number
        hexCount:
          type: integer
        meanAlpha:
          type: number
        maxAlpha:
          type: number
        totalResidentialParcels:
          type: integer
        estimatedDriveways:
          type: integer
        dominantNeighborhood:
          type: string
        label:
          type: string
    CitySummary:
      type: object
      required:
        - city
        - totalHexes
        - hexesAboveThreshold
        - launchZoneCount
        - avgAlphaScore
        - topZoneName
        - topZoneScore
        - topHexH3
        - topHexAlpha
        - topHexLat
        - topHexLng
        - medianDemandScore
        - medianSupplyScore
        - totalEstimatedDriveways
        - algorithmVersion
      properties:
        city:
          type: string
        totalHexes:
          type: integer
        hexesAboveThreshold:
          type: integer
        launchZoneCount:
          type: integer
        avgAlphaScore:
          type: number
        topZoneName:
          type: string
        topZoneScore:
          type: number
        topHexH3:
          type: string
        topHexAlpha:
          type: number
        topHexLat:
          type: number
        topHexLng:
          type: number
        medianDemandScore:
          type: number
        medianSupplyScore:
          type: number
        totalEstimatedDriveways:
          type: integer
        algorithmVersion:
          type: string