# lib/api-zod/src/generated/types/citySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */

export interface CitySummary {
  city: string;
  totalHexes: number;
  hexesAboveThreshold: number;
  launchZoneCount: number;
  avgAlphaScore: number;
  topZoneName: string;
  topZoneScore: number;
  topHexH3: string;
  topHexAlpha: number;
  topHexLat: number;
  topHexLng: number;
  medianDemandScore: number;
  medianSupplyScore: number;
  totalEstimatedDriveways: number;
  algorithmVersion: string;
}