# lib/api-zod/src/generated/types/hexFeature.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { HexFeatureGeometry } from "./hexFeatureGeometry";
import type { HexFeatureProperties } from "./hexFeatureProperties";

export interface HexFeature {
  type: string;
  properties: HexFeatureProperties;
  geometry: HexFeatureGeometry;
}