# lib/api-zod/src/generated/types/hexFeatureProperties.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */

export interface HexFeatureProperties {
  h3Index: string;
  alphaScore: number;
  demandScore: number;
  supplyScore: number;
  competitionPenalty: number;
  neighborhoodName: string;
  launchZoneId: number | null;
  isGoldilocksZone: boolean;
}