# lib/api-zod/src/generated/types/hexGeoJson.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */
import type { HexFeature } from "./hexFeature";

export interface HexGeoJson {
  type: string;
  features: HexFeature[];
}