# lib/api-zod/src/generated/types/launchZone.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */

export interface LaunchZone {
  id: number;
  rank: number;
  centroidLat: number;
  centroidLng: number;
  hexCount: number;
  meanAlpha: number;
  maxAlpha: number;
  totalResidentialParcels: number;
  estimatedDriveways: number;
  dominantNeighborhood: string;
  label: string;
}