# lib/api-zod/src/generated/types/scoredHex.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 0.1.0
 */

export interface ScoredHex {
  h3Index: string;
  lat: number;
  lng: number;
  demandScore: number;
  supplyScore: number;
  competitionPenalty: number;
  alphaScore: number;
  citationCount: number;
  poiCount: number;
  residentialParcelCount: number;
  publicParkingSpots: number;
  estimatedDriveways: number;
  neighborhoodName: string;
  launchZoneId: number | null;
}